chunks: add chunkCount to execution payload

+++ beacon-chain/execution/engine_client.go
@@ -145,7 +145,7 @@ type EngineCaller interface {
 	GetTerminalBlockHash(ctx context.Context, transitionTime uint64) ([]byte, bool, error)
 
 	// Added in Gloas
-	NewBlockHeader(ctx context.Context, payload interfaces.ExecutionData, parentBlockRoot *common.Hash, versionedHashes []common.Hash, executionRequests *pb.ExecutionRequests, chunkCount int) error
+	NewBlockHeader(ctx context.Context, payload interfaces.ExecutionData, parentBlockRoot *common.Hash, versionedHashes []common.Hash, executionRequests *pb.ExecutionRequests, chunkCount uint16) error
 	NewChunkAccessList(ctx context.Context, blockHash common.Hash, chunkIndex uint16, chunkAccessList []byte) error
 	ExecuteChunk(ctx context.Context, blockHash common.Hash, chunk *pb.ExecutionChunk) error
 	FinalizeBlock(ctx context.Context, blockHash common.Hash) ([]byte, error)
